完成物品相关

+++ apps/object/views.py
@@ -15,28 +15,27 @@ class InformationView(APIView):
     def get(self, request):
         data = json.loads(request.body.decode())
         id = data['id']
-        obj = models.LostAndFound.objects.filter(id=id).first()
-        objs_s = LostAndFoundModelSerializer(instance=obj)
-        user = models.Information.objects.filter(user_id=objs_s.data['user_id']).first()
-        objs_s.data['user_name'] = user.name
-        objs_s.data['avator'] = user.avator_url
-        pictures = models.Picture.objects.filter(thing_id=objs_s.data['id'])
+        obj = models.LostAndFound.objects.filter(id=id)
+        objs_s = LostAndFoundModelSerializer(instance=obj, many=True)
+        user = models.Information.objects.filter(user_id=objs_s.data[0]['user']).first()
+        objs_s.data[0]['user_name'] = user.name
+        objs_s.data[0]['avator'] = user.avator_url
+        pictures = models.Picture.objects.filter(thing_id=id)
         picture_data = []
         for picture in pictures:
             picture_data.append(picture.url)
-        objs_s.data['picture'] = picture_data
+        objs_s.data[0]['picture'] = picture_data
         contact = {
             'phone': user.phone,
             'qq': user.qq,
-            'email': user.email,
+            'wechat': user.wechat,
         }
-        objs_s.data['contact'] = contact
-        objs_s.data.pop('type')
-        return JsonResponse({'code': 200, 'message': 'OK', 'data': objs_s.data})
+        objs_s.data[0]['contact'] = contact
+        return JsonResponse({'code': 200, 'message': 'OK', 'data': objs_s.data[0]})
 
     def post(self, request):
         data = json.loads(request.body.decode())
-        data['user_id'] = request.user.id
+        data['user'] = request.user.id
         obj_s = LostAndFoundModelSerializer(data=data)
         if not obj_s.is_valid(raise_exception=True):
             return JsonResponse({'code': 400, 'message': '参数不正确'})
@@ -49,5 +48,8 @@ class InformationDeleteView(APIView):
     def post(self, request):
         data = json.loads(request.body.decode())
         id = data['id']
-        models.LostAndFound.objects.filter(id=id).first().delete()
+        user_id = request.user.id
+        if not models.LostAndFound.objects.filter(id=id, user_id=user_id).exists():
+            return JsonResponse({'code': 404, 'message': '该记录不存在或不是当前登录用户发布'})
+        models.LostAndFound.objects.filter(id=id, user_id=user_id).first().delete()
         return JsonResponse({'code': 200, 'message': 'OK'})
